First commit towards full restructure of the code
Slight modification to adding layer

Basic reformatting of the Layers operations

Time management re-enabled

Fixed some bugs, spliit files that were too big and reformatted up until permalink handling

First commit towards full restructure of the code

Slight modification to adding layer

First fully working version

Fixed typo

First commit towards full restructure of the code

Slight modification to adding layer

Basic reformatting of the Layers operations

Time management re-enabled

Fixed some bugs, spliit files that were too big and reformatted up until permalink handling

First commit towards full restructure of the code

Slight modification to adding layer

First fully working version

Fixed permalink always coloring the map

Fixed typos and scopes
