+++ scripts/wms_sources_configs.json
@@ -1,9 +1,9 @@
 {
-  "weather": {
+  "Weather": {
     "url": "https://geo.weather.gc.ca/geomet",
     "version": "1.3.0"
   },
-  "climate": {
+  "Climate": {
     "url": "https://geo.weather.gc.ca/geomet-climate",
     "version": "1.3.0"
   }
