+++ src/locales/en/common.json
@@ -1,94 +1,72 @@
 {
-  "MSCAnimet": "MSC AniMet",
-  "MSCAnimetEmail": "Weather animation made with MSC AniMet",
-  "AppBarMap": "Map",
-  "AppBarTree": "GeoMet layers",
+  "AnimationFrom": "Weather animation from",
   "AppBarCreate": "Animation configuration",
   "AppBarExport": "Created animation",
-  "AppBarLanguageToggle": "Passer au français",
-  "LayerBarTitle": "List of added GeoMet-Weather Layers",
-  "LayerBarSubtitle": "Controls the layers on the map",
-  "LayerBarOrder": "Layer order",
-  "LayerBarOpacity": "Opacity",
-  "LayerBarLegend": "Legend",
-  "LayerBarRemoveTooltip": "Remove layer",
-  "LayerBarVisibilityTooltip": "Toggle visibility",
-  "LayerBarStartsTooltip": "Start",
-  "LayerBarEndsTooltip": "End",
-  "LayerBarStepTooltip": "Interval",
+  "AppBarMap": "Map",
+  "AppBarTree": "GeoMet layers",
+  "ApplyColor": "Apply color to basemap",
+  "AttributionOSM": "Map by ©OpenStreetMap",
+  "Boundaries": "Canadian geopolitical boundaries",
+  "Climate": "GeoMet-Climate",
+  "Close": "Close",
+  "CMD_WMS": "CMD-WMS",
+  "Colon": ":",
+  "DarkBasemapSwitch": "Dark basemap",
+  "DocumentationURL": "https://eccc-msc.github.io/open-data/msc-animet/readme_en/",
+  "ErrorNoTimeLayer": "Please add a temporal layer",
+  "ExpiredExtentRefreshed": "Some of the selected timesteps are no longer available. The date range slider has been adjusted.",
+  "ExpiredSecondaryLayer": "Timesteps for certain layers are no longer available. Time values have been updated.",
+  "ExtentFullyOOB": "The entire selected temporal extent is no longer available. The date range slider has been reset.",
+  "FPS": "FPS",
+  "FramesPerSecond": "Frames per second",
+  "GeoMetSearchLabel": "Search {wmsSource} layers",
+  "GeoMetWms": "Add {wmsSource} layers",
+  "LayerBarClosestTime": "Closest time: ",
   "LayerBarCurrentTooltip": "Current",
+  "LayerBarEndsTooltip": "End",
   "LayerBarInvisibleTooltip": "Closest date out of bounds. Layer not shown.",
   "LayerBarMapTime": "Current map time: ",
-  "LayerBarClosestTime": "Closest time: ",
-  "SnapLayerToExtent": "Apply temporal extent and interval to animation.",
-  "SnappedLayer": "Temporal extent and interval currently applied to animation.",
-  "MapCustomControls": "Custom controls",
+  "LayerBarOpacity": "Opacity",
+  "LayerBarRemoveTooltip": "Remove layer",
+  "LayerBarStartsTooltip": "Start",
+  "LayerBarStepTooltip": "Interval",
+  "LayerBarVisibilityTooltip": "Toggle visibility",
+  "LayerControlsTitle": "Layers configuration",
+  "LayerStyle": "Select layer style",
+  "Legends": "Legends",
+  "LegendSelector": "Select displayed legend(s)",
+  "MadeWithAniMet": "Made with MSC AniMet",
+  "Major_cities": "Canadian cities",
+  "MapJumpForwardAll": "Jump to the last time step",
+  "MapJumpForwardOne": "Jump to the next time step",
+  "MapPlay": "Play through the time steps at 1 second intervals",
   "MapRewindBackAll": "Move back to the first time step",
   "MapRewindBackOne": "Move back one time step",
-  "MapPlay": "Play through the time steps at 1 second intervals",
-  "MapJumpForwardOne": "Jump to the next time step",
-  "MapJumpForwardAll": "Jump to the last time step",
-  "MP4ExportTitle": "Visualize & export",
-  "MP4ExportSubtitle": "Preview the created animation and download using the button below",
-  "MP4ExportDownload": "Download animation as MP4",
-  "MP4ExportPermalinkTitle": "A copyable link to share the selected layers and extent",
-  "MP4CreateTitle": "Animation configuration",
-  "MP4CreateAddAllLayers": "Show layer titles",
-  "MP4CreateAddTime": "Add time",
-  "MP4CreateAddName": "Add layer name",
-  "framesPerSecond": "Frames per second",
-  "MP4CreateQualitySubheader": "Animation quality (lower is better)",
-  "MP4CreateCancelAnimationCreation": "Cancel",
-  "Mp4Time": "Time",
-  "MP4CreateMap": "Map",
-  "MP4CreateLightMode": "Light mode",
-  "MP4CreateDarkMode": "Dark mode",
-  "DarkBasemapSwitch": "Use dark basemap",
-  "GeoMetWms": "Add {wmsSource} layers",
-  "GeoMetSearchLabel": "Search {wmsSource} layers",
-  "climate": "GeoMet-Climate",
-  "weather": "GeoMet-Weather",
-  "cmd_wms": "CMD-WMS",
-  "MapInteractionsToggle": "Map interactions",
-  "MapControls": "Map controls",
-  "MapGeolocate": "Geolocate",
-  "MP4ProgressBarLabel": "Fetching the data for each layer for all time steps",
-  "MP4CreateProgressLabel": "Stitching all frames into the output MP4 file",
-  "LayerControlsTitle": "Layers configuration",
-  "LocationsSearchTitle": "Preset locations search and visit",
-  "LocationsSearchLabel": "Search for a location you want to jump to",
-  "LocationSearchShort": "Search for a location",
-  "MP4CreateTimeFormat": "Use local time",
-  "MP4CreateCustomTitle": "Animation title (optional)",
+  "ModelRun": "Model run",
   "MP4CreateButtonLabel": "Create animation",
-  "colon": ":",
-  "modelRun": "Model run",
-  "attributionOSM": "Map by ©OpenStreetMap",
-  "madeWithAniMet": "Made with MSC AniMet",
-  "MP4CreateTitleHint": "Limit of 250 characters",
-  "revertTitle": "Revert title",
-  "close": "Close",
+  "MP4CreateCancelAnimationCreation": "Cancel",
+  "MP4CreateCustomTitle": "Animation title (optional)",
+  "MP4CreateDarkMode": "Dark mode",
+  "MP4CreateLightMode": "Light mode",
   "MP4CreateNotifyCancelAnimation": "The animation creation has been cancelled given your browser window has been resized",
-  "animationFrom": "Weather animation from",
-  "socialWeatherHashtags": "weather,opendata",
-  "layerStyle": "Select layer style",
-  "fps": "FPS",
-  "share": "Share",
-  "errorNoTimeLayer": "Please add a temporal layer",
-  "LocalTime": "Local",
-  "timestepsDropdown": "Interval",
-  "legendSelector": "Select displayed legend",
-  "noLegend": "No legend",
-  "documentationURL": "https://eccc-msc.github.io/open-data/msc-animet/readme_en/",
-  "userDoc": "User documentation",
-  "expiredExtentRefreshed": "Some of the selected timesteps are no longer available. The date range slider has been adjusted.",
-  "extentFullyOOB": "The entire selected temporal extent is no longer available. The date range slider has been reset.",
-  "expiredSecondaryLayer": "Timesteps for certain layers are no longer available. Time values have been updated.",
-  "ApplyColor": "Apply color to basemap",
-  "RevertColor": "Revert to original color",
+  "MP4CreateTimeFormat": "Local time",
+  "MP4CreateTitle": "Animation configuration",
+  "MP4CreateTitleHint": "Limit of 250 characters",
+  "MP4ExportDownload": "Download animation as MP4",
+  "MP4ExportSubtitle": "Preview the created animation and download using the button below",
+  "MP4ExportTitle": "Visualize & export",
+  "MSCAnimet": "MSC AniMet",
+  "MSCAnimetEmail": "Weather animation made with MSC AniMet",
   "Overlays": "Overlays",
   "OverlaysTip": "Overlays add information on top of displayed layers.",
-  "Boundaries": "Canadian geopolitical boundaries",
-  "Major_cities": "Canadian cities",
-  "Water_bodies": "Canadian water bodies"
+  "RevertColor": "Revert to original color",
+  "SelectMR": "Select model run",
+  "Share": "Share",
+  "SnapLayerToExtent": "Apply temporal extent and interval to animation.",
+  "SnappedLayer": "Temporal extent and interval currently applied to animation.",
+  "SocialWeatherHashtags": "weather,opendata",
+  "TimestepsDropdown": "Interval",
+  "UserDoc": "User documentation",
+  "Water_bodies": "Canadian water bodies",
+  "Weather": "GeoMet-Weather"
 }
