+++ src/main.js
@@ -4,9 +4,8 @@ import router from "./router";
 import store from "./store";
 import i18n from "./plugins/i18n";
 import vuetify from "./plugins/vuetify";
-import axios from "axios";
 
-Vue.prototype.$http = axios;
+Vue.prototype.$mapLayers = Vue.observable({ arr: [] });
 Vue.config.productionTip = false;
 
 new Vue({
